chore: refactor git module
